+++ libs/react/fpkit/src/index.ts
@@ -1,6 +1,9 @@
-import Badge from './component/badge/fp-badge'
+import Badge from './components/badge/fp-badge'
+import Button from './components/buttons/button'
 
-export { default as FP } from './component/fp'
+export { default as FP } from './components/fp'
+
 export {
-  Badge
+  Badge,
+  Button
 }